Created the Roulette GUI
Was able to get the LayeredPane working. When adding a component to the LayeredPane using the add method, the second argument has to use Integer.valueOf(x) where x is the layer you want the component to be on. Wasn't working before because I was just using the primitive int type.

The spinner.gif file is just a temporary test file setup to go in the center on top of the background table image.

Next test should be to see if buttons work if added on layer 1.  Then we can begin adding all the buttons where they need to go.