Created all the buttons for the board
I stored all the buttons in a array of buttons called buttons.
I also created a function called setButtons that takes in the button array and create and puts all the button in the correct locations. We just need to add the action listeners to the buttons.
I also did a small update to the board.png