I added all the action listeners for the buttons and changed the roulette class a little
I added all the action listeners to all the buttons. When a button is clicked, it adds the bet by calling game.addBet(i, betAmount). I changed how addBet works so it now takes in betAmount. This way we will be able to allow the player to choose a different chip amount.
I also had to move where the components are declared so they are in scope of the actionlistener functions. This allowed me to update the balance jlabel when a bet was placed. I still need to make it so when a button is clicked it will show the picture of the chip on top of it.