Slight Roulette.java restructuring, Started working on action label
I changed up the comments in the Roulette.java file to match the style of the other class file.
I also merged the spin() and calcEarnings() functions and had the spin() function return the number the ball landed on (will probably adjust this later).
Also started working on the label that gets the player feedback on what's going on.