Reorganized more stuff, added new behavior for clearing chips
Removed CustomPanel class.
Edited the look of comments in both files.
Changed indentation for Roulette.java to 2 spaces to maintain consistency with RouletteGUI.java.
Spinning now does not automatically return chips. Instead, the dealer takes the chips that didn't win any bets and the player will now be able to click Clear to return the chips they didn't lose. This is so the player can see which bets won and bet the same again.
Restructured and added to spin() to accomodate a new Boolean bets array.
calcEarnings() was merged with spin().
Changed the purpose of clearChips() to clear losing chips. clearBets() still has the same behavior.